requirements.txt
asyncpg
eth_utils
fastapi
pydantic
python-dotenv
web3
uvicorn